+++ src/classes/Client.php
@@ -2,7 +2,7 @@
 /**
  *	...
  *
- *	Copyright (c) 2014-2024 Christian Würker (ceusmedia.de)
+ *	Copyright (c) 2014-2025 Christian Würker (ceusmedia.de)
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -20,7 +20,7 @@
  *	@category		Tool
  *	@package		CeusMedia.Hymn
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2014-2024 Christian Würker
+ *	@copyright		2014-2025 Christian Würker
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Hymn
  */
@@ -30,7 +30,7 @@
  *	@category		Tool
  *	@package		CeusMedia.Hymn
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2014-2024 Christian Würker
+ *	@copyright		2014-2025 Christian Würker
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Hymn
  *	@todo			code documentation
@@ -46,6 +46,17 @@ class Hymn_Client
 	public const FLAG_NO_FILES			= 64;
 	public const FLAG_NO_INTERACTION	= 128;
 
+	public const FLAGS					= [
+		self::FLAG_VERY_VERBOSE,
+		self::FLAG_VERBOSE,
+		self::FLAG_QUIET,
+		self::FLAG_DRY,
+		self::FLAG_FORCE,
+		self::FLAG_NO_DB,
+		self::FLAG_NO_FILES,
+		self::FLAG_NO_INTERACTION,
+	];
+
 	public const EXIT_ON_END			= 0;
 	public const EXIT_ON_LOAD			= 1;
 	public const EXIT_ON_SETUP			= 2;
@@ -54,13 +65,23 @@ class Hymn_Client
 	public const EXIT_ON_EXEC			= 16;
 	public const EXIT_ON_OUTPUT			= 32;
 
+	public const EXITS					= [
+		self::EXIT_ON_END,
+		self::EXIT_ON_LOAD,
+		self::EXIT_ON_SETUP,
+		self::EXIT_ON_RUN,
+		self::EXIT_ON_INPUT,
+		self::EXIT_ON_EXEC,
+		self::EXIT_ON_OUTPUT,
+	];
+
 	public static string $fileName				= '.hymn';
 
 	public static string $outputMethod			= 'print';
 
 	public static string $language				= 'en';
 
-	public static string $version				= '1.0.2b';
+	public static string $version				= '1.0.2';
 
 	public static string $mode					= 'prod';
 
@@ -75,71 +96,10 @@ class Hymn_Client
 
 	public int $flags							= 0;
 
+	/** @var	?Hymn_Tool_Locale 				$locale			Language support */
 	public ?Hymn_Tool_Locale $locale			= NULL;
 
 	public int $memoryUsageAtStart				= 0;
-
-	protected array $baseArgumentOptions		= [
-		'db'		=> [
-			'pattern'	=> '/^--db=(\S+)$/',
-			'resolve'	=> '\\1',
-			'values'	=> ['yes', 'no', 'only'],
-			'default'	=> 'yes',
-		],
-		'dry'		=> [
-			'pattern'	=> '/^-d|--dry/',
-			'resolve'	=> TRUE,
-			'default'	=> NULL,
-		],
-		'file'		=> [
-			'pattern'	=> '/^--file=(\S+)$/',
-			'resolve'	=> '\\1',
-			'default'	=> '.hymn',
-		],
-		'force'		=> [
-			'pattern'	=> '/^-f|--force$/',
-			'resolve'	=> TRUE,
-			'default'	=> NULL,
-		],
-		'help'		=> [
-			'pattern'	=> '/^-h|--help/',
-			'resolve'	=> TRUE,
-			'default'	=> NULL,
-		],
-		'quiet'		=> [
-			'pattern'	=> '/^-q|--quiet$/',
-			'resolve'	=> TRUE,
-			'default'	=> NULL,
-			'excludes'	=> 'verbose',
-		],
-		'verbose'	=> [
-			'pattern'	=> '/^-v|--verbose$/',
-			'resolve'	=> TRUE,
-			'default'	=> NULL,
-		],
-		'very-verbose'	=> [
-			'pattern'	=> '/^-vv|--very-verbose$/',
-			'resolve'	=> TRUE,
-			'default'	=> NULL,
-			'includes'	=> 'verbose',
-		],
-		'version'	=> [
-			'pattern'	=> '/^--version/',
-			'resolve'	=> TRUE,
-			'default'	=> NULL,
-		],
-		'interactive'	=> [
-			'pattern'	=> '/^--interactive=(\S+)$/',
-			'resolve'	=> '\\1',
-			'values'	=> ['yes', 'no'],
-			'default'	=> 'yes',
-		],
-		'comment'	=> [
-			'pattern'	=> '/^--comment=(\S+)$/',
-			'resolve'	=> '\\1',
-			'default'	=> NULL,
-		]
-	];
 
 	protected static array $commandWithoutConfig	= [
 		//  APP CREATION
@@ -150,6 +110,8 @@ class Hymn_Client
 		'test-syntax',
 		'version',
 	];
+
+	protected array $baseArgumentOptions		= [];
 
 	protected ?Hymn_Structure_Config $config		= NULL;
 
@@ -182,10 +144,13 @@ class Hymn_Client
 		$this->exit					= $exit;
 		$this->originalArguments	= $arguments;
 
+
 		ini_set( 'display_errors', TRUE );
 		error_reporting( E_ALL );
 
 		$phar	= Hymn_Client::$pharPath;
+		$baseArgumentOptionsJson	= file_get_contents( $phar.'baseArgumentOptions.json' );
+		$this->baseArgumentOptions	= json_decode( $baseArgumentOptionsJson, TRUE );
 		if( file_exists( $phar.'.mode' ) )
 			self::$mode		= file_get_contents( $phar.'.mode' ) ?: 'prod';
 		if( file_exists( $phar.'.php' ) )