+++ src/classes/Command/App/Stamp/Diff.php
@@ -2,7 +2,7 @@
 /**
  *	...
  *
- *	Copyright (c) 2017-2024 Christian Würker (ceusmedia.de)
+ *	Copyright (c) 2017-2025 Christian Würker (ceusmedia.de)
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -20,7 +20,7 @@
  *	@category		Tool
  *	@package		CeusMedia.Hymn.Command.App.Stamp
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2017-2024 Christian Würker
+ *	@copyright		2017-2025 Christian Würker
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Hymn
  */
@@ -30,7 +30,7 @@
  *	@category		Tool
  *	@package		CeusMedia.Hymn.Command.App.Stamp
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2017-2024 Christian Würker
+ *	@copyright		2017-2025 Christian Würker
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Hymn
  *	@todo			code documentation
@@ -222,12 +222,12 @@ class Hymn_Command_App_Stamp_Diff extends Hymn_Command_Abstract implements Hymn_
 				$this->client->outVerbose( '--  UPDATE '.strtoupper( $moduleNew->id ).'  --' );
 				$version	= $moduleOld->version;
 				foreach( array_values( $scripts ) as $nr => $script ){
-					$this->client->outVerbose( vsprintf( '--  UPDATE (%d/%d) version %s-> %s', array(
+					$this->client->outVerbose( vsprintf( '--  UPDATE (%d/%d) version %s-> %s', [
 						$nr + 1,
 						count( $scripts ),
-						$version,
-						$script->version
-					) ) );
+						is_object( $version ) ? $version->current : $version,
+						is_object( $script->version ) ? $script->version->current : $script->version
+					] ) );
 					$this->out( trim( $script->sql ) );
 					$version	= $script->version;
 				}