+++ src/classes/Module/Reader2.php
@@ -3,7 +3,7 @@
 /**
  *	Reader for local module XML files.
  *
- *	Copyright (c) 2012-2024 Christian Würker (ceusmedia.de)
+ *	Copyright (c) 2012-2025 Christian Würker (ceusmedia.de)
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -21,7 +21,7 @@
  *	@category		Library
  *	@package		Hymn.Structure.Module
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2012-2024 Christian Würker (ceusmedia.de)
+ *	@copyright		2012-2025 Christian Würker (ceusmedia.de)
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/HydrogenFramework
  */
@@ -44,7 +44,7 @@ use Hymn_Structure_Module_Version as VersionDefinition;
  *	@category		Library
  *	@package		Hymn.Structure.Module
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2012-2024 Christian Würker (ceusmedia.de)
+ *	@copyright		2012-2025 Christian Würker (ceusmedia.de)
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/HydrogenFramework
  */
@@ -129,7 +129,7 @@ class Hymn_Module_Reader2
 		if( !$node->hasAttribute( $attribute ) )
 			return $default ?? FALSE;
 		$value	= $node->getAttribute( $attribute );
-		return in_array( strtolower( $value ), ['true', 'yes', '1', TRUE] );
+		return in_array( strtolower( $value ), ['true', 'yes', 'on', '1'], TRUE );
 	}
 
 	/**