+++ src/classes/Structure/Module.php
@@ -3,7 +3,7 @@
 /**
  *	Module definition.
  *
- *	Copyright (c) 2022-2024 Christian Würker (ceusmedia.de)
+ *	Copyright (c) 2022-2025 Christian Würker (ceusmedia.de)
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -21,7 +21,7 @@
  *	@category		Library
  *	@package		Hymn.Structure.Module
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2022-2024 Christian Würker (ceusmedia.de)
+ *	@copyright		2022-2025 Christian Würker (ceusmedia.de)
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/HydrogenFramework
  */
@@ -32,7 +32,7 @@
  *	@category		Library
  *	@package		Hymn.Structure.Module
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2022-2024 Christian Würker (ceusmedia.de)
+ *	@copyright		2022-2025 Christian Würker (ceusmedia.de)
  *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/HydrogenFramework
  */